Update Calculator.cpp
Calculator V1.0-My First Project in C++

Well, this project is a simple calculator implemented in C++. It is a small part of my learning path because I created this calculator within 4 days of commencing my learning of C++. It indicates my progress while going deeper into programming with this language.

This calculator does the following:

Addition (+)  
Subtraction (-)  
Multiplication (*)  
Division (/)  

**Key Features:**  
**Interactive Input:** The program asks the user to input any one of the operators and two numbers to perform calculation.  
**Structured Output:** Results are displayed in a visually appealing format with ASCII art for a better user experience.  
**Error Handling:** Only processes valid operators.

The above project depicts my understanding with fundamental concepts in C++, such as input and output streams, control structures, and basic arithmetic operations; the above switch and case statements are applied in it.

Feel free to view the source code and provide recommendations for improvement. This is only a stepping stone—considering I am some 30% into mastering C++—to bigger, prolonged programs later on.